creating id for all elements before storing in cosmos (#8486)

+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage/Controllers/InstancesController.cs
@@ -264,18 +264,16 @@ namespace Altinn.Platform.Storage.Controllers
         [Produces("application/json")]
         public async Task<ActionResult<Instance>> Get(int instanceOwnerPartyId, Guid instanceGuid)
         {
-            string instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-
             try
             {
-                Instance result = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+                Instance result = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
                 result.SetPlatformSelfLinks(_storageBaseAndHost);
 
                 return Ok(result);
             }
             catch (Exception e)
             {
-                return NotFound($"Unable to find instance {instanceId}: {e}");
+                return NotFound($"Unable to find instance {instanceOwnerPartyId}/{instanceGuid}: {e}");
             }
         }
 
@@ -377,7 +375,6 @@ namespace Altinn.Platform.Storage.Controllers
                 storedInstance.SetPlatformSelfLinks(_storageBaseAndHost);
 
                 await _partiesWithInstancesClient.SetHasAltinn3Instances(instanceOwnerPartyId);
-
                 return Created(storedInstance.SelfLinks.Platform, storedInstance);
             }
             catch (Exception storageException)
@@ -385,10 +382,13 @@ namespace Altinn.Platform.Storage.Controllers
                 _logger.LogError(storageException, "Unable to create {appId} instance for {instance.InstanceOwner.PartyId}", appId, instance.InstanceOwner.PartyId);
 
                 // compensating action - delete instance
-                await _instanceRepository.Delete(storedInstance);
+                if (storedInstance != null)
+                {
+                    await _instanceRepository.Delete(storedInstance);
+                }
 
-                _logger.LogError("Deleted instance {storedInstance.Id}", storedInstance.Id);
-                return StatusCode(500, $"Unable to create {appId} instance for {instance.InstanceOwner.PartyId} due to {storageException.Message}");
+                _logger.LogError("Deleted instance {storedInstance.Id}", storedInstance?.Id);
+                return StatusCode(500, $"Unable to create {appId} instance for {instance.InstanceOwner?.PartyId} due to {storageException.Message}");
             }
         }
 
@@ -408,15 +408,13 @@ namespace Altinn.Platform.Storage.Controllers
         [Produces("application/json")]
         public async Task<ActionResult<Instance>> Delete(int instanceOwnerPartyId, Guid instanceGuid, [FromQuery] bool hard)
         {
-            string instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-
             Instance instance;
 
-            instance = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+            instance = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
 
             if (instance == null)
             {
-                return NotFound($"Didn't find the object that should be deleted with instanceId={instanceId}");
+                return NotFound($"Didn't find the object that should be deleted with instanceId={instanceOwnerPartyId}/{instanceGuid}");
             }
 
             DateTime now = DateTime.UtcNow;
@@ -474,9 +472,7 @@ namespace Altinn.Platform.Storage.Controllers
             [FromRoute] int instanceOwnerPartyId,
             [FromRoute] Guid instanceGuid)
         {
-            string instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-
-            Instance instance = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+            Instance instance = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
 
             string org = User.GetOrg();
 
@@ -499,7 +495,7 @@ namespace Altinn.Platform.Storage.Controllers
             }
             catch (Exception e)
             {
-                _logger.LogError(e, "Unable to update instance {instanceId}", instanceId);
+                _logger.LogError(e, "Unable to update instance {instanceOwnerPartyId}/{instanceGuid}", instanceOwnerPartyId, instanceGuid);
                 return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
@@ -530,8 +526,7 @@ namespace Altinn.Platform.Storage.Controllers
                 return BadRequest($"Invalid read status: {status}. Accepted types include: {string.Join(", ", Enum.GetNames(typeof(ReadStatus)))}");
             }
 
-            string instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-            Instance instance = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+            Instance instance = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
 
             Instance updatedInstance;
             try
@@ -548,7 +543,7 @@ namespace Altinn.Platform.Storage.Controllers
             }
             catch (Exception e)
             {
-                _logger.LogError(e, "Unable to update read status for instance {instanceId}", instanceId);
+                _logger.LogError(e, "Unable to update read status for instance {instanceOwnerPartyId}/{instanceGuid}", instanceOwnerPartyId, instanceGuid);
                 return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
@@ -579,8 +574,7 @@ namespace Altinn.Platform.Storage.Controllers
                 return BadRequest($"Invalid sub status: {JsonConvert.SerializeObject(substatus)}. Substatus must be defined and include a label.");
             }
 
-            string instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-            Instance instance = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+            Instance instance = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
 
             string org = User.GetOrg();
             if (!instance.Org.Equals(org))
@@ -605,7 +599,7 @@ namespace Altinn.Platform.Storage.Controllers
             }
             catch (Exception e)
             {
-                _logger.LogError(e, "Unable to update sub status for instance {instanceId}", instanceId);
+                _logger.LogError(e, "Unable to update sub status for instance {instaneOwnerPartyId}/{instanceGuid}", instanceOwnerPartyId, instanceGuid);
                 return StatusCode(StatusCodes.Status500InternalServerError);
             }
 
@@ -635,8 +629,7 @@ namespace Altinn.Platform.Storage.Controllers
                 return BadRequest($"Missing parameter value: presentationTexts is misformed or empty");
             }
 
-            string instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-            Instance instance = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+            Instance instance = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
 
             if (instance.PresentationTexts == null)
             {
@@ -682,8 +675,7 @@ namespace Altinn.Platform.Storage.Controllers
                 return BadRequest($"Missing parameter value: dataValues is misformed or empty");
             }
 
-            var instanceId = $"{instanceOwnerPartyId}/{instanceGuid}";
-            Instance instance = await _instanceRepository.GetOne(instanceId, instanceOwnerPartyId);
+            Instance instance = await _instanceRepository.GetOne(instanceOwnerPartyId, instanceGuid);
 
             instance.DataValues ??= new Dictionary<string, string>();
